+++ test/perf/sort/perf.jl
@@ -1,5 +1,6 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
+using Base.GC
 import Base.Sort: QuickSort, MergeSort, InsertionSort
 
 Pkg.add("SortingAlgorithms")